issues #6 Blade ファイルの文法にする ・public/cssフォルダ内に、以前作成したtaichi-portfolioのcssフォルダの中身をコピー＆ペースト→htmlヘッダー部分に記述し呼び出し設定しました。 ・layout.blade.php, gnav.blade.php, header.blade.php, footer.blade.phpを作成し、共通テンプレ化しました。
