+++ resources/views/footer.blade.php
@@ -1,0 +1,6 @@
+<div class="page-top" id="js-page-top">
+        トップへもどる
+      </div>
+      <div class="inner">
+        <div class="copyright">Copyright&copy;TAICHI KOBAYASHI. All Rights Reserved.</div>
+</div>
