+++ resources/views/header.blade.php
@@ -1,0 +1,4 @@
+<h1 class="logo">
+    <div class="logo-title">TAICHI KOBAYASHI</div>
+    <div class="logo-subtitle">PORTFOLIO</div>
+</h1>
