--- resources/views/home/index.blade.php
@@ -1,419 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width,initial-scale=1.0,user-scalable=no">
-  <meta property='og:type' content='website'>
-  <meta property='og:title' content='Webエンジニア TAICHI KOBAYASHI ポートフォリオ'>
-  <meta property='og:url' content='URLが入る'>
-  <meta property='og:description' content='Webエンジニア TAICHI KOBAYASHIのポートフォリオサイトです！'>
-  <meta property="og:image" content="img/ogp.png">
-  <meta name="description" content="Webエンジニア TAICHI KOBAYASHIのポートフォリオサイトです！" />
-  <title>Webエンジニア TAICHI KOBAYASHI ポートフォリオ</title>
-  <link href="https://fonts.googleapis.com/css?family=Open+Sans:700" rel="stylesheet">
-  <link href="css/reset.css" media="all" rel="stylesheet" type="text/css" />
-  <link href="css/photoswipe/photoswipe.css" media="all" rel="stylesheet" type="text/css" />
-  <link href="css/photoswipe/default-skin/default-skin.css" media="all" rel="stylesheet" type="text/css" />
-  <link href="css/style.css" media="all" rel="stylesheet" type="text/css" />
-  <link rel="shortcut icon" href="img/favicon.ico" />
-  <link rel=”canonical” href=”URLが入る” />
-</head>
-
-<body>
-  <div class="wrapper">
-
-    <!-- .gnav -->
-    <nav class="gnav" class="fixed">
-      <ul class="gnav-list">
-        <li class="gnav-item"><a class="gnav-link" href="#works-section">WORKS</a></li>
-        <li class="gnav-item"><a class="gnav-link" href="#feature-section">FEATURE</a></li>
-        <li class="gnav-item"><a class="gnav-link" href="#about-section">ABOUT</a></li>
-        <li class="gnav-item"><a class="gnav-link" href="#skill-section">SKILL</a></li>
-        <li class="gnav-item"><a class="gnav-link" href="#contact-section">CONTACT</a></li>
-      </ul>
-    </nav>
-    <!-- /.gnav -->
-
-    <!-- .header -->
-      <header class="header">
-        <h1 class="logo">
-          <div class="logo-title">TAICHI KOBAYASHI</div>
-          <div class="logo-subtitle">PORTFOLIO</div>
-        </h1>
-      </header>
-    <!-- /.header -->
-
-    <main class="content">
-
-      <!-- .works-section -->
-      <section class="works-section section" id="works-section">
-        <div class="inner">
-          <div class="section-heading">
-            <h2 class="heading-primary">WORKS</h2>
-          </div>
-          <div class="section-lead">
-            <p>Web（デザインとコーディング）制作の記録です。<br>デザイン練習用の架空サイト（ファーストビューのみ）も含んでおります。</p>
-          </div>
-          <div class="section-body works">
-            <div class="works-list js-my-gallery">
-              <figure class="works-item">
-                <a href="img/works/company.jpg" data-size="960x720" class="works-image-wrapper works-link hover-action">
-                  <img class="works-image lazyload fadein" src="img/placeholder4x3.png" data-src="img/works/company.jpg"
-                    alt="" />
-                </a>
-                <figcaption class="works-body">
-                  <div class="works-title">美容系サイト</div>
-                  <div class="works-text">Design / Coding(Responsive) / WordPress</div>
-                  <div class="works-url"><a class="works-url-link" href="https://yahoo.co.jp" target="_blank">Link</a>
-                  </div>
-                </figcaption>
-              </figure>
-              <figure class="works-item">
-                <a href="img/works/hairsalon.jpg" data-size="960x720"
-                  class="works-image-wrapper works-link hover-action">
-                  <img class="works-image lazyload fadein" src="img/placeholder4x3.png"
-                    data-src="img/works/hairsalon.jpg" alt="" />
-                </a>
-                <figcaption class="works-body">
-                  <div class="works-title">建築系サイト</div>
-                  <div class="works-text">Design / Coding(Responsive) </div>
-                  <div class="works-url"><a class="works-url-link" href="https://yahoo.co.jp" target="_blank">Link</a>
-                  </div>
-                </figcaption>
-              </figure>
-              <figure class="works-item">
-                <a href="img/works/girly.jpg" data-size="960x720" class="works-image-wrapper works-link hover-action">
-                  <img class="works-image lazyload fadein" src="img/placeholder4x3.png" data-src="img/works/girly.jpg"
-                    alt="" />
-                </a>
-                <figcaption class="works-body">
-                  <div class="works-title">サイト名が入る</div>
-                  <div class="works-text">Design</div>
-                </figcaption>
-              </figure>
-              <figure class="works-item">
-                <a href="img/works/school.jpg" data-size="960x720" class="works-image-wrapper works-link hover-action">
-                  <img class="works-image lazyload fadein" src="img/placeholder4x3.png" data-src="img/works/school.jpg"
-                    alt="" />
-                </a>
-                <figcaption class="works-body">
-                  <div class="works-title">サイト名が入る</div>
-                  <div class="works-text">Design</div>
-                </figcaption>
-              </figure>
-              <figure class="works-item">
-                <a href="img/works/game.jpg" data-size="960x720" class="works-image-wrapper works-link hover-action">
-                  <img class="works-image lazyload fadein" src="img/placeholder4x3.png" data-src="img/works/game.jpg"
-                    alt="" />
-                </a>
-                <figcaption class="works-body">
-                  <div class="works-title">サイト名が入る</div>
-                  <div class="works-text">Design</div>
-                </figcaption>
-              </figure>
-              <figure class="works-item">
-                <a href="img/works/fashion.jpg" data-size="960x720" class="works-image-wrapper works-link hover-action">
-                  <img class="works-image lazyload fadein" src="img/placeholder4x3.png" data-src="img/works/fashion.jpg"
-                    alt="" />
-                </a>
-                <figcaption class="works-body">
-                  <div class="works-title">サイト名が入る</div>
-                  <div class="works-text">Design</div>
-                </figcaption>
-              </figure>
-            </div>
-          </div>
-        </div>
-      </section>
-      <!-- /.works-section -->
-
-      <!-- .feature-section -->
-      <section class="feature-section section" id="feature-section">
-        <div class="inner">
-          <div class="section-heading">
-            <h2 class="heading-primary">FEATURE</h2>
-          </div>
-          <div class="section-lead">
-            <p>
-              １サイトの解説
-            </p>
-          </div>
-          <div class="section-body feature">
-            <div class="feature-item">
-              <div class="feature-image-wrapper">
-                <a href="https://www.yahoo.co.jp/" target="_blank" class="hover-action"><img clas="feature-image"
-                    src="img/feature/feature.jpg" alt=""></a>
-              </div>
-              <div class="feature-body">
-                <p>
-                  <span class="text-bold">サイト名：</span><br>サイト名が入る<br><a href="#" target="_blank">https://●●●.com</a></p>
-                <p><span class="text-bold">担当：</span><br>Design / Coding (Responsive) / WordPress / Writing</p>
-                <p><span
-                    class="text-bold">コメント：</span><br>テキストテキストテキストテキストテキストテキストテキストテキストテキスト<br>テキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキスト
-                </p>
-                <p>
-                  テキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキスト</p>
-                <p>テキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキスト</p>
-              </div>
-            </div>
-          </div>
-        </div>
-      </section>
-      <!-- /.feature-section -->
-
-      <!-- .about-section -->
-      <section class="about-section section" id="about-section">
-        <div class="inner">
-          <div class="section-heading">
-            <h2 class="heading-primary">ABOUT</h2>
-          </div>
-          <div class="section-lead">
-            <p>
-              こんにちは！
-            </p>
-          </div>
-          <div class="section-body about">
-            <div class="about-item">
-              <div class="about-image-wrapper">
-                <img class="about-image lazyload fadein" data-src="img/about/profile.jpg" alt="">
-              </div>
-              <div class="about-body">
-                <p>
-                  テキストテキストテキストテキストテキストテキストテキストテキストテキスト<br>テキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキスト
-                </p>
-                <p>
-                  テキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキスト>テキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキスト
-                </p>
-              </div>
-            </div>
-          </div>
-        </div>
-      </section>
-      <!-- /.about-section -->
-
-      <!-- .skill-section -->
-      <section class="skill-section section" id="skill-section">
-        <div class="inner">
-          <div class="section-heading">
-            <h2 class="heading-primary">SKILL</h2>
-          </div>
-          <div class="section-lead">
-            <p>
-              未経験ですが頑張ります！
-            </p>
-          </div>
-          <div class="section-body skill">
-            <div class="skill-list">
-              <div class="skill-item">
-                <div class="skill-image-wrapper">
-                  <img class="skill-image" src="img/skill/icon_html.png" alt="">
-                </div>
-                <div class="skill-body">
-                  <h3 class="skill-title">HTML5/CSS3</h3>
-                  <div class="skill-rating rating">
-                    <div class="rate rate4"></div>
-                  </div>
-                  <p>
-                    テキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキスト</p>
-                </div>
-              </div>
-              <div class="skill-item">
-                <div class="skill-image-wrapper">
-                  <img class="skill-image" src="img/skill/icon_js.png" alt="">
-                </div>
-                <div class="skill-body">
-                  <h3 class="skill-title">JavaScript</h3>
-                  <div class="skill-rating rating">
-                    <div class="rate rate3"></div>
-                  </div>
-                  <p>
-                    テキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキスト</p>
-                </div>
-              </div>
-              <div class="skill-item">
-                <div class="skill-image-wrapper">
-                  <img class="skill-image" src="img/skill/icon_wp.png" alt="">
-                </div>
-                <div class="skill-body">
-                  <h3 class="skill-title">WordPress</h3>
-                  <div class="skill-rating rating">
-                    <div class="rate rate4"></div>
-                  </div>
-                  <p>
-                    テキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキスト</p>
-                </div>
-              </div>
-              <div class="skill-item">
-                <div class="skill-image-wrapper">
-                  <img class="skill-image" src="img/skill/icon_ps.png" alt="">
-                </div>
-                <div class="skill-body">
-                  <h3 class="skill-title">Photoshop</h3>
-                  <div class="skill-rating rating">
-                    <div class="rate rate4"></div>
-                  </div>
-                  <p>
-                    テキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキスト</p>
-                </div>
-              </div>
-              <div class="skill-item">
-                <div class="skill-image-wrapper">
-                  <img class="skill-image" src="img/skill/icon_ai.png" alt="">
-                </div>
-                <div class="skill-body">
-                  <h3 class="skill-title">Illustrator</h3>
-                  <div class="skill-rating rating">
-                    <div class="rate rate3"></div>
-                  </div>
-                  <p>
-                    テキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキスト</p>
-                </div>
-              </div>
-              <div class="skill-item">
-                <div class="skill-image-wrapper">
-                  <img class="skill-image" src="img/skill/icon_xd.png" alt="">
-                </div>
-                <div class="skill-body">
-                  <h3 class="skill-title">XD</h3>
-                  <div class="skill-rating rating">
-                    <div class="rate rate3"></div>
-                  </div>
-                  <p>
-                    テキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキスト</p>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </section>
-      <!-- /.skill-section -->
-
-      <!-- .contact-section -->
-      <section class="contact-section section" id="contact-section">
-        <div class="inner">
-          <div class="section-heading">
-            <h2 class="heading-primary">CONTACT</h2>
-          </div>
-          <div class="section-lead">
-            <p>
-              お問い合わせは、<br class="sp-only">下記お問い合わせフォームからお願いいたします。
-            </p>
-          </div>
-          <div class="section-body contact">
-            <form method="post" action="">
-                <label for="name">お名前</label>
-                <input type="text" name="name" value="" id="name">
-                  <label for="email">メールアドレス</label>
-                <input type="text" name="email" value="" id="email">
-                <label for="message">メッセージ</label>
-                <textarea name="message" rows="5" cols="70" id="message"></textarea>
-              <input type="submit" value="送信">
-            </form>
-            
-            <!-- <ul class="contact-list">
-              <li class="contact-item">
-                <a class="contact-link" href="https://twitter.com/" target="_blank"><img class="contact-image"
-                    src="img/contact/icon_tw.png" alt=""></a>
-              </li>
-              <li class="contact-item">
-                <a class="contact-link" href="https://www.facebook.com/" target="_blank"><img class="contact-image"
-                    src="img/contact/icon_fb.png" alt=""></a>
-              </li>
-            </ul>
-            <div class="contact-text">higoronookonaigadaiji@gmail.com</div> -->
-          </div>
-        </div>
-      </section>
-      <!-- /.contact-section -->
-
-    </main>
-
-    <!-- .footer -->
-    <footer class="footer">
-      <div class="page-top" id="js-page-top">
-        トップへもどる
-      </div>
-      <div class="inner">
-        <div class="copyright">Copyright&copy;TAICHI KOBAYASHI. All Rights Reserved.</div>
-      </div>
-    </footer>
-    <!-- /.footer -->
-
-  </div>
-
-  <!-- 下記はPhotoswipeプラグイン用 -->
-  <!-- Root element of PhotoSwipe. Must have class pswp. -->
-  <div class="pswp" tabindex="-1" role="dialog" aria-hidden="true">
-
-    <!-- Background of PhotoSwipe.
-       It's a separate element as animating opacity is faster than rgba(). -->
-    <div class="pswp__bg"></div>
-
-    <!-- Slides wrapper with overflow:hidden. -->
-    <div class="pswp__scroll-wrap">
-
-      <!-- Container that holds slides.
-          PhotoSwipe keeps only 3 of them in the DOM to save memory.
-          Don't modify these 3 pswp__item elements, data is added later on. -->
-      <div class="pswp__container">
-        <div class="pswp__item"></div>
-        <div class="pswp__item"></div>
-        <div class="pswp__item"></div>
-      </div>
-
-      <!-- Default (PhotoSwipeUI_Default) interface on top of sliding area. Can be changed. -->
-      <div class="pswp__ui pswp__ui--hidden">
-
-        <div class="pswp__top-bar">
-
-          <!--  Controls are self-explanatory. Order can be changed. -->
-
-          <div class="pswp__counter"></div>
-
-          <button class="pswp__button pswp__button--close" title="Close (Esc)"></button>
-
-          <button class="pswp__button pswp__button--share" title="Share"></button>
-
-          <button class="pswp__button pswp__button--fs" title="Toggle fullscreen"></button>
-
-          <button class="pswp__button pswp__button--zoom" title="Zoom in/out"></button>
-
-          <!-- Preloader demo https://codepen.io/dimsemenov/pen/yyBWoR -->
-          <!-- element will get class pswp__preloader--active when preloader is running -->
-          <div class="pswp__preloader">
-            <div class="pswp__preloader__icn">
-              <div class="pswp__preloader__cut">
-                <div class="pswp__preloader__donut"></div>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        <div class="pswp__share-modal pswp__share-modal--hidden pswp__single-tap">
-          <div class="pswp__share-tooltip"></div>
-        </div>
-
-        <button class="pswp__button pswp__button--arrow--left" title="Previous (arrow left)">
-        </button>
-
-        <button class="pswp__button pswp__button--arrow--right" title="Next (arrow right)">
-        </button>
-
-        <div class="pswp__caption">
-          <div class="pswp__caption__center"></div>
-        </div>
-
-      </div>
-
-    </div>
-
-  </div>
-  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-  <script type='text/javascript' src="js/photoswipe/photoswipe.min.js"></script>
-  <script type='text/javascript' src="js/photoswipe/photoswipe-ui-default.js"></script>
-  <script type='text/javascript' src="js/photoswipe/photoswipe-setup.js"></script>
-  <script type='text/javascript' src="js/lazysizes/lazysizes.min.js"></script>
-  <script type='text/javascript' src="js/script.js"></script>
-</body>
-</html>
-
