--- resources/views/layouts/master_insert.blade.php
@@ -1,45 +1,0 @@
-<!DOCTYPE HTML>
-    <html lang="ja">
-        <head>
-            <meta charset="utf-8">
-            <meta http-equiv="X-UA-Compatible" content="IE=edge">
-            <meta name="viewport" content="width=device-width, initial-scale=1">
-            <title>@yield('title')</title>
-            <!-- Bootstrap -->
-            <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
-
-            <!--[if lt IE 9]>
-                <script src="https://oss.maxcdn.com/html5shiv/3.7.2/html5shiv.min.js"></script>
-                <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
-            <![endif]-->
-            @yield('styles')
-
-            <!-- FontAwesome -->
-            <link href="http://netdna.bootstrapcdn.com/font-awesome/4.6.3/css/font-awesome.css" rel="stylesheet">
-        </head>
-
-        <body>
-            <nav class="navbar navbar-default navbar-fixed-top" role="navigation">
-                <div class="container-fluid">
-                    <div class="navbar-header">
-                        <a class="navbar-brand" href="#">@yield('title')</a>
-                    </div>
-                </div>
-            </nav>
-
-            <div class="container" style="margin-top: 40px;">
-                @yield('content')
-            </div>
-
-            <footer class="fixed-bottom footer">
-                <div class="container">
-                    <span class="text-muted">@yield('title')</span>
-                </div>
-            </footer>
-
-            <!-- jQuery -->
-            <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.3/jquery.min.js"></script>
-            <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.4/js/bootstrap.min.js"></script>
-            @yield('script')
-        </body>
-    </html>
